State: Move isRequestingJetpackJumpstartStatus selector to global selector directory
